requirements[1].txt
fastapi
uvicorn
openai-whisper
torch
requests
